+++ Userland/Libraries/LibJS/Runtime/Intl/DateTimeFormatConstructor.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021, Tim Flynn <trflynn89@serenityos.org>
+ * Copyright (c) 2021-2022, Tim Flynn <trflynn89@serenityos.org>
  *
  * SPDX-License-Identifier: BSD-2-Clause
  */
@@ -26,5 +26,7 @@ private:
 
     JS_DECLARE_NATIVE_FUNCTION(supported_locales_of);
 };
+
+ThrowCompletionOr<DateTimeFormat*> initialize_date_time_format(GlobalObject& global_object, DateTimeFormat& date_time_format, Value locales_value, Value options_value);
 
 }
